FIX None issue
