+++ tests/formatted_test_data_1.proto
@@ -3,7 +3,7 @@ asdddddddd
 price demo  API request message
 sdfghjk
 */
-syntax = "proto3";  // The price sdfgfhg API request message //asd /*asdfghj*/阿斯蒂芬
+syntax = "proto3";  // The price sdfgfhg API request message //asd /*asdfghj*/ AS
 
 /*
 asdddddddd
