+++ tests/formatted_test_data_1_align_with_4_indents_all_top_comments.proto
@@ -2,7 +2,7 @@
 asdddddddd
 price demo  API request message
 sdfghjk
-The price sdfgfhg API request message //asd /*asdfghj*/阿斯蒂芬
+The price sdfgfhg API request message //asd /*asdfghj*/ AS
 */
 syntax = "proto3";
