+++ tests/test_data_1.proto
@@ -3,7 +3,7 @@
    price demo  API request message
     */
    // sdfghjk
-    syntax = "proto3";//The price sdfgfhg API request message //asd /*asdfghj*/
+    syntax = "proto3";//The price sdfgfhg API request message //asd /*asdfghj*/ AS
    /*
    asdddddddd
     */
