+++ tests/test_parse.py
@@ -3,49 +3,6 @@ from proto_formatter.parser import ProtoParser
 from tests.helper import read_file, write_file
 
 
-def make_test_data():
-    c = [
-        {
-            'file_name': 'formatted_test_data_1.proto',
-            'indents': 2,
-            'all_top_comments': False,
-            'equal_sign': False
-        },
-        {
-            'file_name': 'formatted_test_data_1_align_with_equal_sign.proto',
-            'indents': 2,
-            'all_top_comments': False,
-            'equal_sign': True
-        },
-        {
-            'file_name': 'formatted_test_data_1_align_with_equal_sign_4_indents.proto',
-            'indents': 4,
-            'all_top_comments': False,
-            'equal_sign': True
-        },
-        {
-            'file_name': 'formatted_test_data_1_align_with_4_indents_all_top_comments.proto',
-            'indents': 4,
-            'all_top_comments': True,
-            'equal_sign': False
-        },
-        {
-            'file_name': 'formatted_test_data_1_align_with_equal_sign_4_indents_all_top_comments.proto',
-            'indents': 4,
-            'all_top_comments': True,
-            'equal_sign': True
-        }
-    ]
-    parser = ProtoParser()
-    lines = parser.read_lines('test_data_1.proto')
-    protobuf_obj = parser.parse(lines)
-
-    for e in c:
-        actual_text = Formatter(indents=e['indents'], all_top_comments=e['all_top_comments'],
-                                equal_sign=e['equal_sign']).to_string(protobuf_obj)
-        write_file(e['file_name'], actual_text)
-
-
 def test_parse_align():
     expected_text = read_file('formatted_test_data_1.proto')
     parser = ProtoParser()
@@ -99,5 +56,3 @@ def test_parse_align_with_equal_sign_4_indents_all_top_comments():
     actual_text = Formatter(equal_sign=True, indents=4, all_top_comments=True).to_string(protobuf_obj)
 
     assert expected_text == actual_text
-
-make_test_data()
